% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnOneDS.R
\name{returnOneDS}
\alias{returnOneDS}
\title{Returns 1}
\usage{
returnOneDS()
}
\value{
1
}
\description{
This function returns the number one.
}
\details{
This function simply returns the integer value 1
}
\author{
Ayisha
}
